# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sensor Management API Documentation",
        "description": "API Documentation for Sensor Management System",
        "contact": {
            "name": "",
            "email": ""
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/sensor-readings/{sensor}": {
            "get": {
                "tags": [
                    "Sensor Readings"
                ],
                "summary": "Get latest sensor reading",
                "description": "Returns the latest reading for a specific sensor",
                "operationId": "getLatestSensorReading",
                "parameters": [
                    {
                        "name": "sensor",
                        "in": "path",
                        "description": "Sensor identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SensorReadingResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sensor not found"
                    }
                }
            }
        },
        "/sensor-readings/collection/{sensor}": {
            "get": {
                "tags": [
                    "Sensor Readings"
                ],
                "summary": "Get sensor readings collection",
                "description": "Returns a collection of readings for a specific sensor within a time range, limited to 40 points",
                "operationId": "getSensorReadingsCollection",
                "parameters": [
                    {
                        "name": "sensor",
                        "in": "path",
                        "description": "Sensor identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Start date for readings (defaults to 1970-01-01)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "End date for readings (defaults to current time)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SensorReadingResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sensor not found"
                    }
                }
            }
        },
        "/sensor-readings/collection/{sensor}/raw": {
            "get": {
                "tags": [
                    "Sensor Readings"
                ],
                "summary": "Get raw sensor readings data",
                "description": "Returns raw paginated data for a specific sensor within a time range",
                "operationId": "getRawSensorReadings",
                "parameters": [
                    {
                        "name": "sensor",
                        "in": "path",
                        "description": "Sensor identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Start date for readings",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "End date for readings",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "value",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SensorReadingResource"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string"
                                                },
                                                "last": {
                                                    "type": "string"
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sensor not found"
                    }
                }
            }
        },
        "/sensor-readings/collection/{sensor}/download": {
            "get": {
                "tags": [
                    "Sensor Readings"
                ],
                "summary": "Download sensor readings as CSV",
                "description": "Generates and downloads a CSV file with readings for a specific sensor within a time range",
                "operationId": "downloadSensorReadingsCsv",
                "parameters": [
                    {
                        "name": "sensor",
                        "in": "path",
                        "description": "Sensor identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Start date for readings",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "End date for readings",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date-time"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CSV file download",
                        "content": {
                            "text/csv": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Sensor not found"
                    }
                }
            }
        },
        "/sensor-readings": {
            "post": {
                "tags": [
                    "Sensor Readings"
                ],
                "summary": "Create a new sensor reading",
                "description": "Records a new reading for a specific sensor",
                "operationId": "createSensorReading",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SensorReadingCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sensor reading successfully recorded!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Missing or invalid admin key"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sensorAdminKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SensorReadingCreateRequest": {
                "required": [
                    "sensor_name",
                    "value"
                ],
                "properties": {
                    "sensor_name": {
                        "description": "Name of the sensor",
                        "type": "string",
                        "maxLength": 50
                    },
                    "value": {
                        "description": "Reading value",
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
...
            "SensorReadingResource": {
                "properties": {
                    "value": {
                        "description": "Sensor reading value",
                        "type": "number",
                        "format": "float"
                    },
                    "symbol": {
                        "description": "Unit of measurement symbol",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Timestamp when reading was recorded",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ValidationErrorResponse": {
                "properties": {
                    "type": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sensorAdminKey": {
                "type": "apiKey",
                "name": "X-Admin-Token",
                "in": "header"
            }
        }
    },
...
}